move e2e helpers into e2e_util package
