+++ tests/e2e/tests/e2e_util/harness/__init__.py
@@ -1,0 +1,17 @@
+#
+# Copyright 2023 Canonical, Ltd.
+#
+from e2e_util.harness.base import Harness, HarnessError
+from e2e_util.harness.juju import JujuHarness
+from e2e_util.harness.local import LocalHarness
+from e2e_util.harness.lxd import LXDHarness
+from e2e_util.harness.multipass import MultipassHarness
+
+__all__ = [
+    HarnessError,
+    Harness,
+    JujuHarness,
+    LocalHarness,
+    LXDHarness,
+    MultipassHarness,
+]
