--- tests/e2e/tests/harness.py
@@ -1,8 +1,0 @@
-#
-# Copyright 2023 Canonical, Ltd.
-#
-from harness_base import Harness
-from harness_juju import JujuHarness
-from harness_local import LocalHarness
-from harness_lxd import LXDHarness
-from harness_multipass import MultipassHarness
